refactor(time): move period into the time namespace
