+++ codecov.yml
@@ -91,13 +91,13 @@ fixes:
   - github.com/go-fries/fries/mysql/canal/positioner/redis/v4::mysql/canal/positioner/redis/
   - github.com/go-fries/fries/mysql/canal/server/v4::mysql/canal/server/
   - github.com/go-fries/fries/otel/otlp/v4::otel/otlp/
-  - github.com/go-fries/fries/period/v4::period/
   - github.com/go-fries/fries/poll/v4::poll/
   - github.com/go-fries/fries/ptr/v4::ptr/
   - github.com/go-fries/fries/recovery/v4::recovery/
   - github.com/go-fries/fries/retry/v4::retry/
   - github.com/go-fries/fries/signal/v4::signal/
   - github.com/go-fries/fries/support/v4::support/
+  - github.com/go-fries/fries/time/period/v4::time/period/
   - github.com/go-fries/fries/timezone/v4::timezone/
   - github.com/go-fries/fries/udp/v4::udp/
   - github.com/go-fries/fries/webhook/v4::webhook/
