+++ time/period/doc.go
