+++ time/period/errors.go
