+++ versions.yaml
@@ -138,9 +138,6 @@ module-sets:
       # Otel
       - github.com/go-fries/fries/otel/otlp/v4
 
-      # Period
-      - github.com/go-fries/fries/period/v4
-
       # Poll
       - github.com/go-fries/fries/poll/v4
 
@@ -164,6 +161,9 @@ module-sets:
 
       # Support
       - github.com/go-fries/fries/support/v4
+
+      # Time
+      - github.com/go-fries/fries/time/period/v4
 
       # Timezone
       - github.com/go-fries/fries/timezone/v4
